Item Pointed at Home Dropdown Fix
When a menu link at the 1st level pointed to home and had dropdown
items, the dropdown wasn't displayed. This was because of an issue in
the better_mm_get_children function, which called
menu_link_get_preferred method to fetch the path of the home link for
building the menu tree. However, the function returned a non-existant
item from the navigation menu, which resulted in no dropdown. This was
fixed by using menu_tree_all_data to fetch the menu structure manually
from the database, and return the children.