Get Menu Children Changed
The issues #20 and #21 stem from the use of the code that retrieves menu
item children (also amended in the last commit). The code only checked
for a menu item's path, and then got the menu item's children using
menu_link_get_preferred. I have entirely restructured this code which
now fixes these issues. However, a break could potentially happen if
there are two menu items with the exact same title, and same path. This
should not be done though, so it is something that I am deeming
acceptable, but can be changed later if that functionality is needed for
some reason.